lim 분활 정리
