+++ accounts/forms.py
@@ -1,5 +1,6 @@
 from django import forms
-from .models import user, Spend, Income
+from .models import user
+from calendars.models import Spend, Income
 
 class LemonSignupForm(forms.Form):
     uid = forms.CharField()
@@ -14,13 +15,13 @@ class LemonSignupForm(forms.Form):
 
 
 
-class SpendForm(forms.ModelForm):
-    class Meta:
-        model = Spend
-        fields = ['user','kind','spend_date','amount','place', 'way', 'category', 'card', 'memo']
-
-
-class IncomeForm(forms.ModelForm):
-    class Meta:
-        model = Income
-        fields = ['user','kind','income_date','amount','income_way','memo']
+# class SpendForm(forms.ModelForm):
+#     class Meta:
+#         model = Spend
+#         fields = ['user','kind','spend_date','amount','place', 'way', 'category', 'card', 'memo']
+#
+#
+# class IncomeForm(forms.ModelForm):
+#     class Meta:
+#         model = Income
+#         fields = ['user','kind','income_date','amount','income_way','memo']
