+++ accounts/templates/base.html
@@ -9,7 +9,7 @@
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.18.1/moment.min.js"></script>
     <script src="//code.jquery.com/jquery-3.3.1.min.js"></script>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">      <script src="http://netdna.bootstrapcdn.com/bootstrap/3.0.2/js/bootstrap.min.js" type="text/javascript"></script> 
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">      <script src="http://netdna.bootstrapcdn.com/bootstrap/3.0.2/js/bootstrap.min.js" type="text/javascript"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
     <link rel = "stylesheet" href="{% static 'css/style.css'%}">
     <link rel="stylesheet" href="{% static 'css/swiper.css'%}">
@@ -27,7 +27,7 @@
 
 
 
-    
+
 
 
 <style>
@@ -82,7 +82,7 @@
         width: 100%;
         height: 4.5rem;
         background: -webkit-linear-gradient(to right, #31BC8A, #4C8AE5);  /* Chrome 10-25, Safari 5.1-6 */
-        background: linear-gradient(to right, #31BC8A, #4C8AE5); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */    
+        background: linear-gradient(to right, #31BC8A, #4C8AE5); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
         position: fixed;
         top: 0;
         left: 0;
@@ -165,7 +165,7 @@
         <div id="footer">
         {% block footer %}
         <ul class="bottom-nav-wrap">
-            <li class="bottom-nav-calendar" style="width: 25%; height:100%; text-align:center;"><a href="{% url 'accounts:calendar' %}"><i class="far fa-calendar-alt"></i><br>가계부</a></li>
+            <li class="bottom-nav-calendar" style="width: 25%; height:100%; text-align:center;"><a href="{% url 'calendars:calendar' %}"><i class="far fa-calendar-alt"></i><br>가계부</a></li>
             <li class="bottom-nav-stock" style="width: 25%; height:100%; text-align:center;"><a href="{% url 'accounts:stock' %}"><i class="fas fa-chart-area"></i><br>주식</a></li>
             <li class="bottom-nav-addlist" style="width: 25%; height:100%; text-align:center;"><a href="{% url 'accounts:addlist' %}"><i class="fas fa-shopping-cart"></i><br>구매내역</a></li>
             <li class="bottom-nav-myinfo" style="width: 25%; height:100%; text-align:center;"><a href="{% url 'accounts:myinfo' %}"><i class="fas fa-user"></i><br>내정보</a></li>
@@ -178,17 +178,17 @@
 <script>
     document.documentElement.addEventListener('touchstart', function (event) {
         if (event.touches.length > 1) {
-            event.preventDefault(); 
-        } 
+            event.preventDefault();
+        }
     }, false);
 
-var lastTouchEnd = 0; 
+var lastTouchEnd = 0;
 
 document.documentElement.addEventListener('touchend', function (event) {
     var now = (new Date()).getTime();
     if (now - lastTouchEnd <= 300) {
-            event.preventDefault(); 
-        } lastTouchEnd = now; 
+            event.preventDefault();
+        } lastTouchEnd = now;
     }, false);
 </script>
 </html>
