+++ lemon/urls.py
@@ -29,6 +29,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('accounts.urls')),
     path('account/', include('allauth.urls')),
+    path('',include('calendars.urls')),
     path('.*', TemplateView.as_view(template_name='index.html')),
 ]
